requirements.txt
streamlit
langchain
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-community
faiss-cpu
pypdf
openai
google-generativeai
anthropic
sentence-transformers
transformers
torch
huggingface_hub
requests
arxiv
pytesseract
pdf2image
python-magic
tabulate
nltk
pandas
numpy
Pillow
python-multipart
